+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/State/Query.hs
@@ -18,11 +18,19 @@ module Cardano.Ledger.Api.State.Query (
 
   -- * @GetFilteredDelegationsAndRewardAccounts@
   module Cardano.Ledger.Api.State.Query.StakeDelegation,
+
+  -- * UTxO queries
+  module Cardano.Ledger.Api.State.Query.UTxO,
+
+  -- * Debug queries
+  module Cardano.Ledger.Api.State.Query.Debug,
 ) where
 
+import Cardano.Ledger.Api.State.Query.Debug
 import Cardano.Ledger.Api.State.Query.Epoch
 import Cardano.Ledger.Api.State.Query.Governance
 import Cardano.Ledger.Api.State.Query.PParams
 import Cardano.Ledger.Api.State.Query.Pool
 import Cardano.Ledger.Api.State.Query.Snapshot
 import Cardano.Ledger.Api.State.Query.StakeDelegation
+import Cardano.Ledger.Api.State.Query.UTxO